build-sys: remove DEFAULT_INCLUDES workaround
The automake stuff uses "-I.". as a default gcc option for
includes. This is a problem for source code where is local includes
with a same name like system includes (e.g.  mntent.h,
paths.h). Possible workaround is overwrite the automake
DEFAULT_INCLUDES variable. But this solution produces warnings. The
best way (this patch) is probably rename the files and remove
DEFAULT_INCLUDES.

[modified for mount.nilfs2.c and umount.nilfs2.c]

Signed-off-by: Karel Zak <kzak@redhat.com>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>