nilfs2-utils: rename member prefix of struct nilfs_suinfo
NILFS2 module failed to build for IA64 machines due to a side-effect
of shortname alias for the struct seginfo, which is
defined in linux-2.6.xx/include/asm-ia64/siginfo.h.

We've renamed the member prefix of struct nilfs_suinfo to avoid
the failure.  Since the struct is included in nilfs_fs.h, this
change should be reflected also to the utility package.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>
Acked-by: Koji Sato <sato.koji@lab.ntt.co.jp>