nilfs2-utils: decide to add or modify mtab from actual entries
The previous implementation decided to add or modify mtab
by whether a mount operation is remount or not.
This is problematic when applying fake or nomtab options,
so this changes it to decide by the actual mtab entries;
if an entry with a same device and a same mount point has
found, the existing mtab entry will be modified instead of
adding a new entry.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>