nilfs2-utils: add minor flag for checkpoint created by internal operation
Nilfs creates checkpoints even for garbage collection or metadata
updates such as checkpoint mode change.  So, user often sees
checkpoints created only by such internal operations.

This is inconvenient in some situations.  For example, an application
which changes checkpoints to snapshots, will fall into an infinite
loop because it cannot distinguish those checkpoints.

To resolve this sort of problem, a new flag was added to checkpoint
for identification.  This allows userland tools to use the flag.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>