nilfs2-utils: fix privilege escalation vulnerability in mkfs.nilfs2
Eric Sandeen gave me a report that mkfs.nilfs2 has a crucial security
problem that allows privilege escalation.

When '-c' option is specified, mkfs.nilfs2 calls system() to invoke
badblocks command against a given device in command line.  On the
otherhand, the make file of mkfs.nilfs2 is written to set a suid bit
on it after installation; thus, it has a risk that arbitrary scripts
embedded in the device string can be executed with an elevated
privilege.

This patch fix the security issue by dropping the suid bit from the
make file.

Thanks to Eric Sandeen and Steve Grubb for pointing out this problem.

Reported-by: Steve Grubb <sgrubb@redhat.com>
Reported-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>