nilfs2-utils: allow optional unit suffix for min/max clean segments
This allows the following style declaration for min_clean_segments and
max_clean_segments parameters of the cleanerd conf file
(i.e. /etc/nilfs_cleanerd.conf):

 min_clean_segments  8MB

 max_clean_segments  10%

When '%' is used the value is calculated as percentage for the total
number of segments.

The unit suffix is optional and the parameter represents number of
segments if no unit suffix is given.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>