mount.nilfs2: add "uhelper=" option
Standard mount command takes uhelper=<helper_name> option to tell the mount
is under <helper_name> control.  When this option is specified,
umount.<helper_name> is called on unmount time.

This option is used when mount is triggered by udisks. udisks is a dbus
service daemon.  The service is used by Palimpsest.  To controll nilfs2
partition from palimpsest, interpret the option is required.

Signed-off-by: Jiro SEKIBA <jir@unicus.jp>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>