Compare device number rather than device name for mount check.
To check whether a file system is mounted, nilfs-tune was comparing
the device name to the contents of /etc/mtab.  This check does not
detect a mounted file system when the device has multiple names, which
is a common case when using LVM.  (LVM creates symbolic links from
/dev/mapper/vg-lv to /dev/dm-X.)

This patch, borrowed from the analogous check in e2fsprogs, compares
the device number (field st_rdev of struct stat) rather than the
device name.

Signed-off-by: Matteo Frigo <athena@fftw.org>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>