nilfs-utils: add block count option to lscp command
This adds "-b" (or "--show-block-count") option and collaterally "-g"
(or "--show-increment") option to the lscp command to allow switching
amount of used blocks and that of appended blocks.

Note that the value of the block count field is not yet valid.  It
needs a change of the kernel code that adds block count support.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>