lscp: hide internal checkpoints and add all option to show them
This patch slightly changes behavior of lscp command.

The current lscp command displays all checkpoints including minor
checkpoints, which are created as the result of garbage collection or
snapshot operations.  (Each minor checkpoint is accompanied by an "i"
flag in the lscp).

This patch adds all option to lscp, and changes it so as to show all
checkpoints when the all option is specified and does not show the
minor checkpoint by default.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>