man nilfs-clean
I believe I found a typo in the nilfs-clean man page. (is an interger
value -> is an integer value)
I cloned the git repo, edited the file and then I did a git diff.
Not sure if this is the proper way to post a "patch" but here's my
first try.

Signed-off-by: Tiago Caxias <tiago.caxias@yahoo.ca>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>