libnilfsgc: fix broken live/dead determination function
This is a long-standing bug fix.

The current live/dead determination function of nilfs virtual block,
nilfs_vdesc_is_live() is broken.

It may return wrong result if duration period of the virtual block is
included in the range from the minimum snapshot number to the maximum
snapshot number.

This fixes it.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>