nilfs_cleanerd: try to shrink protected segments when hit them
nilfs_cleanerd may stop garbage collection even if it has room for
reclaming more segments.  This issue is reproducible with the
following steps:

 - make snapshots
 - change all snapshots back to checkpoints.
 - run GC with nilfs-clean -p1
 - remove all data from the current file system tree.
 - run GC again with nilfs-clean -p1

In this situation, cleanerd hits segments protected by the file
system; one of log cursors written in super blocks was not updated and
prevents garbage collection.

This fixes the issue by trying to update the log cursor when GC failed
due to the protected segments.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>