mount.nilfs2: fix wrong error handling in do_mount_one routine
mount.nilfs2 tries to restart cleanerd for rw->ro remount and rw->rw
remount if a mount system call was failed.

However, it does not update mtab properly and outputs an error message
"mount.nilfs2: failed to restart nilfs_cleanerd" even though it
successfully restarted the cleanerd.

This was caused by a bug of return code handling for
nilfs_launch_cleanerd() function.

This fixes it.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>