nilfs-utils: mkfs.nilfs2 should check presence of NILFS2 volume on device
The mkfs.nilfs2 utility should check presence of existing file system
or partition table on device and to warn a user about possibility to
destroy data by mkfs activity. This patch uses libblkid library for
detection of presence of file system or partition table on opened
device. If libblkid detects any known signature then mkfs.nilfs2
informs a user about potential danger to destroy existing file system
or partition table. The execution of mkfs.nilfs2 stops with offering
to make decision about continuation or abortion of operation.

Moreover, this patch adds "-f" option that gives opportunity to force
overwrite when an existing file system is detected on the device. By
default, mkfs.nilfs2 will not write to the device if it suspects that
there is a file system on the device already. The man page of
mkfs.nilfs2 was modified by description of "-f" option.

Reported-by: Hendrik Levsen <hendrik@levsen.org>
Signed-off-by: Vyacheslav Dubeyko <slava@dubeyko.com>
Cc: Martin Steigerwald <Martin@lichtvoll.de>
Tested-by: Vyacheslav Dubeyko <slava@dubeyko.com>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>