mkfs.nilfs2: erase first and second sectors unless boot sector exists
There is a filesystem which writes a super block on the first sector
of the device.  One of such filesystems is xfs.  Since the current
mkfs.nilfs2 does not erase the first and the second sectors, blkid can
detect that the partition is used with the old file system even after
mkfs.nilfs2 has formatted the device.

This fixes the issue by erasing the first and the second if boot
sector does not exist on the first sector.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>