cleanerd: call _exit(2) twice for ensuring not being a session leader
Current daemonize() function of cleanerd call _exit(2) only once
during its process of becoming a daemon process.  But in the linux
environment, a daemon should call _exit(2) twice for ensuring not
being a session leader.  If a process don't do that, unexpected SIGHUP
can be sent to the process (though it happens rarely). The signal
would be confusing event for cleanerd of nilfs.  This patch removes
this potential problem.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>