cleanerd: output error message of oom_adjust() with syslog()
Error messages of oom_adjust() is ignored since it is called after
daemonize() and daemonize() closes stdin, stdout, and stderr.  This
fixes the issue by directing these error messages to syslog.

This also modifies the error handling of oom_adjust() so that cleanerd
doesn't abort even if oom_adjust() failed.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>