{mount,umount}.nilfs2: use libmount by default
Previously, build with libmount was optional but is enabled
conditionally depending on whether /etc/mtab is a symlink or not.

Nowadays, /etc/mtab is a symlink to /proc/self/mounts or so in most
distros.  So, we let mount/umount helpers of nilfs2 use libmount by
default and revise the README file to reflect the change.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>