libmount, libblkid: use PKG_CHECK_MODULES for probing
After a round to add this to buildroot, I learned that
the proper way to check the dependencies seems to point to pkg-config.
This commit migrates the libmount & libblkid checks to use
PKG_CHECK_MODULES.
I added a similar patch to current buildroot, and this succeeds
all daily tests as well.

[ ryusuke: added fallback to configure.ac ]

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>