nilfs-utils: do not use AC_TYPE_SIGNAL and RETSIGTYPE
The AC_TYPE_SIGNAL macro is diagnosed as obsolete in Autoconf 2.70 and
later:

 configure.ac:164: warning: The macro `AC_TYPE_SIGNAL' is obsolete.
 configure.ac:164: You should run autoupdate.
 ./lib/autoconf/types.m4:776: AC_TYPE_SIGNAL is expanded from...
 configure.ac:164: the top level

AC_TYPE_SIGNAL defines the value of RETSIGTYPE macro, which will be
set to 'void' or 'int' depending on platforms.  But, in the first
place, signal handlers returning 'int' became obsolete long ago and no
modern C platforms use the legacy return type.

So, this replaces RETSIGTYPE with 'void' in codebase of nilfs-utils in
addition to removing the AC_TYPE_SIGNAL macro.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>