mkfs, nilfs-tune: fix gcc8 warnings for volume name buffers
For the volume name buffer in super blocks, we store a volume name
without a terminating null character if it is truncated to the buffer
size.

This suppresses false positive -Wstringop-truncation warnings of GCC 8
related to the volume name buffers and adds comments so as to clarify
the treatment of truncation on them.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>