nilfs-resize: fix abnormal message output when shrinking file system
Although the return value of nilfs_resize_count_inuse_segments() is type
ssize_t, nilfs_shrink_online(), which calls it, receives the return value
as a local variable of type __u64.

This causes the following abnormality:

1. If nilfs_resize_count_inuse_segments() returns an error, it is not
   treated as an error, and the message output assuming its success it not
   skipped.
2. An abnormal progress message like "Moving xx in-use segment(null)." is
   output during file system shrink in a 32-bit environment due to type
   mismatch.

Fix these issues by changing the type of the local variable in question to
ssize_t.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>