mount_libmount: fix umount failure after remount
For nilfs2 mounts that are read-only mounted and then remounted in
read/write mode, nilfs_cleanerd will no longer be shut down by umount or
remount, and the file system will not be able to be unmounted unless the
user manually kills it.

This is because the mount root key "ROOT", which libmount needs to
identify the mount entry on utab that stores file system-specific mount
attributes, is not set by remount or fake mount.  This prevents
libmount-based mount and umount programs from correctly inheriting
cleanerd's PID, making it impossible to control cleanerd when remounting
or unmounting.

Fix this issue by searching the mount table and completing the missing
mount root with the mount entry information obtained when performing
remount or fake mount.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>