Make installation of kernel uapi header files optional
The bundled kernel's uapi header files ("nilfs2_ondisk.h" and
"nilfs2_api.h") are included in kernel-related development packages
in modern Linux distributions, resulting in duplicate installations.

In order to eliminate this duplication, make the installation of these
files optional and not install them by default.

The utilities will continue to refer to the included uapi header files,
and will only allow installation on the system, which can be specified
using the "--enable-uapi-header-install" option in configure file
arguments.

This is a workaround for older environments and should normally be
disabled.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>