nilfs.h: move byte order conversion macros to compat.h
Macros such as cpu_to_le32()/le32_to_cpu() were originally intended
for use within the kernel, not for user space.

If should not exist in "nilfs.h", which is intended for use in user
space, but is only for compatibility, so move it to "compat.h".

Also, type definitions for endianness checks such as __le32 can now be
assumed to exist in "linux/types.h", so delete them.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>