lib/sb.c: fix super block read failure from disk image file
Unlike mkfs.nilfs2, nilfs-tune and dumpseg commands will cause errors
if used directly on a disk image files.  This is because
__nilfs_sb_read(), which reads superblocks, uses the BLKGETSIZE64 ioctl
command to get the device size, which fails for disk image files.

Fix this issue by using the size information obtained from the fstat
system call for regular files.

[ resolved conflicts and build issues with v2.2.y ]

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>