nilfs-tune: fix the exit code when mount is detected
Since the exit code when the mount is detected is the success value,
it is not possible to determine whether the program was successfully
executed or not just from the exit code.

Fix this issue by returning an error exit code when a mount is
detected on the device during the mount check performed without
the "-f" option.

Also use errx() to output the error message and adjust the message
format.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>