libmountchk: determine if image file is mounted via a loop device
The check_mount() function used to check the mount of a block device
does not support determining the mount status when the source is an
image file mounted via a loop device.

As a result, nilfs-tune can rewrite the loop device mount source file
without the force option "-f".

Fix this issue by making check_mount() detect if a loop device is
attached and a file system is mounted on it.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>