nilfs-resize: fix abnormal truncation when specifying size with fraction
When we specify a fractional device size for nilfs-resize command, the
device size will be abnormally truncated due to an error in size
correction calculation (missing type cast).

For example, if we specify 8,796,093,023,200 bytes (= 8TiB - 32 bytes)
as the device size for an 8TiB partition, the size will be adjusted to
512 bytes as shown below, resulting in an insufficient size error:

 $ sudo nilfs-resize /dev/loop0 8796093023200
 size 8796093023200 is not aligned to sector size. truncated to 512.
 Partition size = 8796093023232 bytes.
 Shrink the filesystem size from 8796093023232 bytes to 512 bytes.
 Error: the filesystem does not have enough free space.
        At least 10 more segments (83886080 bytes) are required.
 Aborted.

Fix this issue by inserting the missing typecast in the alignment
correction calculation.  After applying this modification, the device
size will be rounded down appropriately to the sector size as shown
below:

 $ sudo nilfs-resize /dev/loop0 8796093023200
 size 8796093023200 is not aligned to sector size. truncated to 8796093022720.
 Partition size = 8796093023232 bytes.
 Shrink the filesystem size from 8796093023232 bytes to 8796093022720 bytes.
 No segments will be truncated.
 Do you wish to proceed (y/N)? y
 Done.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>