libnilfs: nilfs_open: identify the mount point device
When nilfs_open() opens a mount point to obtain an ioctl FD, it may be
overlaid by another mount and may be a directory of a file system
mounted with a different device.

Most nilfs utilities cannot operate in situations where the mount
point is hidden, but with the current implementation, an ioctl FD is
opened for the overlying file system, and ioctl commands are issued
wrongly for it.

Fix this issue by detecting such condition and returning a ENOENT error
in that case.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>