nilfs-resize: make the progress bar more accurate
nilfs_resize_count_inuse_segments() is used to count the segments to
be moved in order to evaluate the progress of the progress bar.
However, the current implementation counts segments whose flags are in
the reclaimable state (dirty, not active, not error), and does not
count active segments, resulting in a discrepancy.

To represent the progress bar more acculately, change the count to
literally "in-use" (dirty) segments instead of reclaimable segments.

Also, there is some confusion regarding the term "in-use" in the
description of the nilfs_resize_reclaim() function, so correct that
as well.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>