nilfs-resize: rename function nilfs_resize_inuse_segments()
nilfs_resize_inuse_segments() checks the sufile's segusage flags and
collects and returns the numbers of segments that are reclaimable
(dirty, not active, not error).  However, this function name is a bit
confusing, since active segments are also "in-use".

Change the function name to the more direct
nilfs_resize_reclaimable_segments() in order to treat "inuse" as meaning
that it is literally in use, including active segments.

[ resolved conflicts for v2.2.y ]

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>