bin/*: correct error message for invalid long option specification
The commands under "bin/" output the error messages by themselves by
setting opterr to 0, and are compatible with getopt_long, but if
an invalid long option is specified, the option that causes the error
will not be displayed as shown below:

 lscp: invalid option --

Also lscp and lssu do not properly display error messages for options
that require an additional argument and it is not provided.

Fix these issues by not setting opterr to 0 and leaving error output
for argument options to getopt() or getopt_long()'s built-in features.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>