mkfs.nilfs2: do not use own type blocknr_t
The block address is now represented by uint64_t, but only mkfs.nilfs2
uses its own type "blocknr_t".

Eliminate the own type so that the bit width of variables that handle
it can be made stricter (such as when displaying).

Additionally, to avoid obscuring the meaning of "start" in identifiers
intended to be block addresses, rename them to "start_blocknr".

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>