build: fix build failure due to conflicting dirs after branch switching
Since Git does not automatically remove untracked directories when
switching branches, a name that was a directory in v2.2.y can persist
and block the creation of a binary with the same name in master.

This change automates the cleanup within the build process, ensuring
a robust build even when switching between versions without manual
'git clean -xdf'.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>