build: support generating pkg-config information for shared libraries
Add support for generating pkg-config files, which contain build
dependency information for shared libraries, and add a brief
explanation of the changes in the README file.

The location where the pkg-config files are installed is
auto-detected, but you can change the install location or disable the
generation of the pkg-config files alltogether using the
--with-pkgconfigdir option. (This can also be disabled with the
--without-pkgconfigdir option.)

This can help external tools use the libraries, but currently it is
intended as a convenience for external development tools to easily
accommodate differences in build environments.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>