build: decouple pkg-config detection from pc file installation
Currently, the detection of the pkg-config tool is coupled with the
configuration for installing .pc files.  This prevents using
pkg-config for checking dependencies if the user chooses not to
install .pc files (using --without-pkgconfigdir).

So, move the PKG_PROG_PKG_CONFIG check to the top level, allowing
pkg-config to be used for dependency resolution regardless of the
installation target settings.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>