README: add pkg-config to list of required packages
Add 'pkg-config' to the list of required packages in the README file.

The build system now supports detection of external libraries via
'.pc' files and generation of the project's '.pc' files.  Although a
fallback mechanism is implemented, making 'pkg-config' not strictly
mandatory, it is required to generate the '.pc' files and to resolve
build dependencies through them.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>